ssm summary, externalreferences
