+++ src/packages/@ncigdc/modern_components/SsmExternalReferences/SsmExternalReferences.js
@@ -1,7 +1,6 @@
 // @flow
 
 import React from 'react';
-import Relay from 'react-relay/classic';
 import { compose, withPropsOnChange } from 'recompose';
 import externalReferenceLinks from '@ncigdc/utils/externalReferenceLinks';
 import EntityPageVerticalTable from '@ncigdc/components/EntityPageVerticalTable';
@@ -20,22 +19,24 @@ const styles = {
   },
 };
 
-type TProps = {
-  node: Object,
-  dbSNP: string,
-};
-
-const SsmExternalReferencesComponent = compose(
-  withPropsOnChange(['node'], ({ node }) => ({
-    dbSNP: node.consequence.hits.edges.reduce(
-      (acc, c) =>
-        c.node.transcript.annotation.dbsnp_rs
-          ? c.node.transcript.annotation.dbsnp_rs
-          : acc,
-      '',
-    ),
-  })),
-)(({ node, dbSNP }: TProps = {}) =>
+export default compose(
+  withPropsOnChange(
+    ['viewer'],
+    ({ viewer: { explore: { ssms: { edges } } } }) => {
+      const node = edges[0].node;
+      return {
+        dbSNP: node.consequence.hits.edges.reduce(
+          (acc, c) =>
+            c.node.transcript.annotation.dbsnp_rs
+              ? c.node.transcript.annotation.dbsnp_rs
+              : acc,
+          '',
+        ),
+        node,
+      };
+    },
+  ),
+)(({ node, dbSNP } = {}) =>
   <EntityPageVerticalTable
     title={
       <span>
@@ -72,47 +73,3 @@ const SsmExternalReferencesComponent = compose(
     style={{ ...styles.summary, ...styles.column, alignSelf: 'flex-start' }}
   />,
 );
-
-export const SsmExternalReferencesQuery = {
-  initialVariables: {
-    withDbsnp_rs: {
-      op: 'AND',
-      content: [
-        {
-          op: 'NOT',
-          content: {
-            field: 'consequence.transcript.annotation.dbsnp_rs',
-            value: 'MISSING',
-          },
-        },
-      ],
-    },
-  },
-  fragments: {
-    node: () => Relay.QL`
-      fragment on Ssm {
-        cosmic_id
-        consequence {
-          hits(first: 1 filters: $withDbsnp_rs) {
-            edges {
-              node {
-                transcript {
-                  annotation {
-                    dbsnp_rs
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    `,
-  },
-};
-
-const SsmExternalReferences = Relay.createContainer(
-  SsmExternalReferencesComponent,
-  SsmExternalReferencesQuery,
-);
-
-export default SsmExternalReferences;
