+++ src/packages/@ncigdc/modern_components/SsmExternalReferences/index.js
@@ -1,0 +1,3 @@
+import Component from './SsmExternalReferences';
+import createRenderer from './SsmExternalReferences.relay';
+export default createRenderer(Component);
