+++ src/packages/@ncigdc/modern_components/SsmSummary/index.js
@@ -1,0 +1,3 @@
+import Component from './SsmSummary';
+import createRenderer from './SsmSummary.relay';
+export default createRenderer(Component);
